+++ clustered/client/src/main/java/org/ehcache/clustered/client/internal/store/StrongServerStoreProxy.java
@@ -16,8 +16,8 @@
 package org.ehcache.clustered.client.internal.store;
 
 import org.ehcache.clustered.client.internal.EhcacheClientEntity;
-import org.ehcache.clustered.common.messages.EhcacheEntityResponse;
-import org.ehcache.clustered.common.messages.ServerStoreMessageFactory;
+import org.ehcache.clustered.common.internal.messages.EhcacheEntityResponse;
+import org.ehcache.clustered.common.internal.messages.ServerStoreMessageFactory;
 import org.ehcache.clustered.common.store.Chain;
 import org.ehcache.core.spi.function.NullaryFunction;
 import org.slf4j.Logger;
