+++ clustered/client/src/test/java/org/ehcache/clustered/client/internal/EhcacheClientEntityFactoryTest.java
@@ -20,7 +20,7 @@ import java.util.UUID;
 
 import org.ehcache.clustered.client.internal.service.ClusteredStoreManagerConfigurationException;
 import org.ehcache.clustered.common.ServerSideConfiguration;
-import org.ehcache.clustered.lock.common.LockMessaging.HoldType;
+import org.ehcache.clustered.common.internal.lock.LockMessaging.HoldType;
 import org.ehcache.clustered.lock.client.VoltronReadWriteLockClient;
 import org.junit.Test;
 import org.terracotta.connection.Connection;
