+++ clustered/clustered-dist/build.gradle
@@ -28,6 +28,11 @@ ext {
   docsFolder = "$buildDir/docs/asciidoc"
 }
 
+dependencies {
+  compile project(':clustered:client')
+  compile project(':clustered:common')
+}
+
 apply plugin: 'distribution'
 apply plugin: EhDistribute
 
@@ -37,8 +42,6 @@ configurations {
 }
 
 dependencies {
-  compile project(':clustered:client')
-  compile project(':clustered:common')
   compile "org.terracotta.internal:client-runtime:$terracottaCoreVersion" changing true
   compile "org.terracotta.internal:client-logging:$terracottaCoreVersion" changing true
