+++ clustered/common/src/main/java/org/ehcache/clustered/common/internal/messages/EhcacheEntityResponse.java
@@ -13,16 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehcache.clustered.common.messages;
+
+package org.ehcache.clustered.common.internal.messages;
 
 
 import org.ehcache.clustered.common.store.Chain;
 import org.terracotta.entity.EntityResponse;
 
-/**
- *
- * @author cdennis
- */
 public abstract class EhcacheEntityResponse implements EntityResponse {
 
   public enum Type {
