+++ clustered/common/src/main/java/org/ehcache/clustered/common/internal/messages/EhcacheEntityResponseFactory.java
@@ -13,22 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehcache.clustered.common.messages;
 
-import org.terracotta.entity.EntityMessage;
+package org.ehcache.clustered.common.internal.messages;
 
-/**
- * Defines messages for interactions with an {@code EhcacheActiveEntity}.
- */
-public abstract class EhcacheEntityMessage implements EntityMessage {
+import org.ehcache.clustered.common.store.Chain;
 
-  public enum Type {
-    SERVER_STORE_OP,
-    LIFECYCLE_OP;
+public class EhcacheEntityResponseFactory {
+
+  public EhcacheEntityResponse success() {
+    return EhcacheEntityResponse.Success.INSTANCE;
   }
 
-  public abstract Type getType();
+  public EhcacheEntityResponse failure(Exception cause) {
+    return new EhcacheEntityResponse.Failure(cause);
+  }
 
-  public abstract byte getOpCode();
-
+  public EhcacheEntityResponse response(Chain chain) {
+    return new EhcacheEntityResponse.GetResponse(chain);
+  }
 }
