--- clustered/common/src/test/java/org/ehcache/clustered/common/messages/Util.java
@@ -1,42 +1,0 @@
-/*
- * Copyright Terracotta, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.ehcache.clustered.common.messages;
-
-import org.ehcache.clustered.common.store.Chain;
-import org.ehcache.clustered.common.store.Element;
-
-import java.util.Iterator;
-
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
-import static org.ehcache.clustered.common.store.Util.readPayLoad;
-
-/**
- *
- */
-public final class Util {
-
-  private Util() {
-  }
-
-  public static void assertChainHas(Chain chain, long... payLoads) {
-    Iterator<Element> elements = chain.iterator();
-    for (long payLoad : payLoads) {
-      assertThat(readPayLoad(elements.next().getPayload()), is(Long.valueOf(payLoad)));
-    }
-    assertThat(elements.hasNext(), is(false));
-  }
-}
