+++ clustered/server/src/main/java/org/ehcache/clustered/server/EhcacheServerEntityService.java
@@ -15,9 +15,9 @@
  */
 package org.ehcache.clustered.server;
 
-import org.ehcache.clustered.common.messages.EhcacheCodec;
-import org.ehcache.clustered.common.messages.EhcacheEntityMessage;
-import org.ehcache.clustered.common.messages.EhcacheEntityResponse;
+import org.ehcache.clustered.common.internal.messages.EhcacheCodec;
+import org.ehcache.clustered.common.internal.messages.EhcacheEntityMessage;
+import org.ehcache.clustered.common.internal.messages.EhcacheEntityResponse;
 import org.terracotta.entity.ConcurrencyStrategy;
 import org.terracotta.entity.MessageCodec;
 import org.terracotta.entity.PassiveServerEntity;
